Add manifest count per component metric
